fix- created some hooks
