+++ .gitignore
@@ -23,10 +23,4 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 /test-results/
-/playwright/.cache/
 /playwright-report/
-/blob-report/
-.github/workflows/playwright.yml
-e2e/
-tests-examples/
-tests/
