+++ src/components/ui/badge/Badge.tsx
@@ -15,6 +15,7 @@ const badge = cva(
         red: ["bg-[#ffebe8]", "text-[#d1242f]", "border-[#ffcac8]"],
       },
       size: {
+        extraSmall: ["text-[10px]", "px-2", "py-0.5"],
         small: ["text-[10px]", "px-2.5", "py-0.5"],
         medium: ["text-base", "px-4", "py-1.5"],
       },
